requirements.txt
Django==1.5.5
South==0.8.2
Werkzeug==0.9.4
django-crispy-forms==1.4.0
django-debug-toolbar==0.10.2
django-extensions==1.2.5
django-geojson==2.1.1
-e git+https://github.com/rosscdh/django-jsonfield.git@a08821100651130cc9f58ccd22ffb71c2658eabd#egg=django_jsonfield-dev
django-leaflet==0.8.1
django-parsley==0.1
-e git+https://github.com/rosscdh/django-templatetag-handlebars.git@e32fb140f862f7931b7a360ffcc0cbe637f6d913#egg=django_templatetag_handlebars-dev
djangorestframework==2.3.8
gunicorn==18.0
psycopg2==2.5.1
six==1.4.1
sqlparse==0.1.9
wsgiref==0.1.2
